fixed issues in routing
