+++ src/routes/index.js
@@ -5,14 +5,15 @@ const winston = require('winston');
 const path = require('path');
 const express = require('express');
 
+const { set } = require('grunt');
 const meta = require('../meta');
 const controllers = require('../controllers');
 const controllerHelpers = require('../controllers/helpers');
 const plugins = require('../plugins');
-const bugReportController = require('../controllers/bugReports');
 const authRoutes = require('./authentication');
 const writeRoutes = require('./write');
 const helpers = require('./helpers');
+const reportBugController = require('../controllers/reportBug');
 
 const { setupPageRoute } = helpers;
 
@@ -80,6 +81,7 @@ _mounts.categories = (app, name, middleware, controllers) => {
 	setupPageRoute(app, '/recent', [], controllers.recent.get);
 	setupPageRoute(app, '/top', [], controllers.top.get);
 	setupPageRoute(app, '/unread', [middleware.ensureLoggedIn], controllers.unread.get);
+	setupPageRoute(app, '/report-bug', [], reportBugController.getBugReportForm);
 };
 
 _mounts.category = (app, name, middleware, controllers) => {
@@ -219,7 +221,4 @@ function addRemountableRoutes(app, router, middleware, mounts) {
 
 		_mounts[original](router, mount, middleware, controllers);
 	});
-}
-_mounts.reportBug = (app, middleware, controllers) => {
-	setupPageRoute(app, '/report-bug', [], controllers.bugReports.getBugReportForm);
 };
